% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_experiment.R
\name{design_experiment}
\alias{design_experiment}
\title{Set Parameters for In Silico gRNA Experiment}
\usage{
design_experiment(
  ncellpops = 1,
  edge_probs = 0.2,
  ntfs_per_cellpop = sample(10:20, ncellpops, replace = TRUE),
  grna_to_ctrl_ratio = runif(1),
  nhk = 9000,
  negenes = 500,
  nsimulations = 100,
  ngrnas_per_target = 3,
  ncells_per_model = 1000,
  census_interval = 10,
  tau = 100/3600,
  ntargets = sum(ntfs_per_cellpop),
  ncells_in_experiment = 10000,
  target_tf_only = TRUE,
  grna_library = "default",
  experiment_type = "ko",
  ctrl_label = "CTRL",
  cache_dir = file.path(getwd(), "cache")
)
}
\arguments{
\item{ncellpops}{null}

\item{edge_probs}{null}

\item{ntfs_per_cellpop}{null}

\item{nhk}{null}

\item{nsimulations}{null}

\item{ngrnas_per_target}{null}

\item{ncells_per_model}{null}

\item{census_interval}{null}

\item{tau}{null}

\item{target_tf_only}{null}

\item{grna_library}{null}

\item{experiment_type}{null}

\item{ctrl_label}{null}

\item{cache_dir}{null}

\item{ngenes}{null}

\item{ntarget}{null}
}
\description{
Set Parameters for In Silico gRNA Experiment
}
