% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fantom5}
\alias{fantom5}
\title{Gene Regulatory Networks from the Fantom5 Database}
\format{
A list of adjacency matrix. Each tissue type avaliable has it's own graph.
}
\usage{
fantom5
}
\description{
Gene Regulatory Networks from the Fantom5 Database
}
\keyword{datasets}
