% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grna_lib_meta}
\alias{grna_lib_meta}
\title{gRNA Library On-Target and Off-Target PDFs}
\format{
A list of lists. Each individual sub list is a grna library. for each library you have 2 dataframes containing on and off target pdfs.
}
\usage{
grna_lib_meta
}
\description{
gRNA Library On-Target and Off-Target PDFs
}
\keyword{datasets}
